Tipo de i modificado para int

+++ aula-01/java/com/github/vnduda/ufg/cs/aula01/Ex13.java
@@ -2,8 +2,7 @@ package com.github.vnduda.ufg.cs.aula01;
 
 public class Ex13 {
 
-    // TODO verifique se o tipo de i deve ser float (parece inadequado)
-    public static float raiz(float n, float i) {
+    public static float raiz(float n, int i) {
 
         if (n < 0) {
             throw new IllegalArgumentException("Intervalo da variável n inválido");
